move connection related commands to a new trait and object (for underlying messages), add echo and ping
